[Change] Animate out snackbar message faster
When snackbar visibility goes from true to false to true,
the rendering of the message happens faster than the
ReactTransitionGroup animations.

Animating out the snackbar message with a duration of 0
hides the flash of the next message.

However, it would be better to render the next message
after the entire out-animation is complete. (Maybe do
the out-animation in componentWillUpdate.)