[Fix] Don't decide what tags can be searched
This fixes the error where a newly attached tag cannot
become a search tag because it's not yet listed in the
search suggestions, which updates right before opening.

Checking if a search term is in the list of suggestions
was trying to prevent the situation of zero search results.
Firstly, that situation is acceptable. Secondly, it alone
did not prevent that situation, which also arises after
deleting all files of search result.