[Fix] Clear tagger.suggestions.requestsPending when tagger value is empty       and check requestsPending before setting state from database response
This fixes the delay in showing suggestions after attaching a tag. After
attaching a tag, tagger value is cleared. When tagger value is empty,
there is no need to go to the database for suggestions, so there is no
need to wait for any pending database responses.

When a tag is attached, the tagger value is cleared. If there was an